poc/OneSDKPOC/requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
OneSDKPOC==1.0.0
pycparser==2.21
pyOpenSSL==23.1.1
python-dotenv==1.0.0
requests==2.30.0
urllib3==1.26.7
